% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{class_intervals}
\alias{class_intervals}
\title{Class Intervals Used By A griddy Object}
\usage{
class_intervals(x, ...)
}
\arguments{
\item{x}{A griddy result object.}

\item{...}{Reserved for future methods.}
}
\value{
A tibble of class intervals when available.
}
\description{
Class Intervals Used By A griddy Object
}
