% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griddy-package.R
\docType{package}
\name{griddy-package}
\alias{griddy}
\alias{griddy-package}
\title{griddy: Geospatial Distribution Dynamics}
\description{
Tidy helpers for exploratory geospatial distribution dynamics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dshkol/griddy}
  \item \url{https://dshkol.github.io/griddy/}
  \item Report bugs at \url{https://github.com/dshkol/griddy/issues}
}

}
\author{
\strong{Maintainer}: Dmitry Shkolnik \email{shkolnikd@gmail.com}

}
\keyword{internal}
