% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rank_mobility}
\alias{plot_rank_mobility}
\title{Plot Rank Mobility}
\usage{
plot_rank_mobility(x)
}
\arguments{
\item{x}{A \code{grd_rank_mobility} object returned by \code{\link[=rank_mobility]{rank_mobility()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Rank Mobility
}
