% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_transition_matrix}
\alias{plot_transition_matrix}
\title{Plot A Transition Matrix}
\usage{
plot_transition_matrix(x)
}
\arguments{
\item{x}{A \code{grd_markov} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot A Transition Matrix
}
