% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{steady_state}
\alias{steady_state}
\title{Estimate Stationary Distribution}
\usage{
steady_state(x)
}
\arguments{
\item{x}{A \code{grd_markov} object or transition probability matrix.}
}
\value{
A numeric vector.
}
\description{
Estimate Stationary Distribution
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
prob <- matrix(c(0.8, 0.2, 0.3, 0.7), nrow = 2, byrow = TRUE)
steady_state(prob)
\dontshow{\}) # examplesIf}
}
