% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{transition_matrix}
\alias{transition_matrix}
\title{Extract A Transition Matrix}
\usage{
transition_matrix(x, type = c("probability", "count"), lag_class = NULL)
}
\arguments{
\item{x}{A \code{grd_markov} or \code{grd_spatial_markov} object.}

\item{type}{\code{"probability"} or \code{"count"}.}

\item{lag_class}{Optional lag class for \code{grd_spatial_markov}.}
}
\value{
A matrix.
}
\description{
Extract A Transition Matrix
}
