% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{lag_intervals}
\alias{lag_intervals}
\title{Spatial Lag Intervals Used By A Spatial Markov Object}
\usage{
lag_intervals(x)
}
\arguments{
\item{x}{A \code{grd_spatial_markov} object.}
}
\value{
A tibble of spatial-lag class intervals.
}
\description{
Spatial Lag Intervals Used By A Spatial Markov Object
}
