% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_spatial_markov}
\alias{plot_spatial_markov}
\title{Plot Spatial Markov Matrices}
\usage{
plot_spatial_markov(x)
}
\arguments{
\item{x}{A \code{grd_spatial_markov} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Spatial Markov Matrices
}
