% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_transition_matrix}
\alias{plot_transition_matrix}
\title{Plot A Transition Matrix}
\usage{
plot_transition_matrix(x)
}
\arguments{
\item{x}{A \code{grd_markov} object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot A Transition Matrix
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
panel <- data.frame(id = rep(1:4, each = 2), year = rep(2020:2021, 4), value = 1:8)
markov <- panel |>
  classify_dynamics(id, year, value, k = 2) |>
  markov_dynamics(id, year, class)

plot_transition_matrix(markov)
\dontshow{\}) # examplesIf}
}
