% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{steady_state}
\alias{steady_state}
\title{Estimate Stationary Distribution}
\usage{
steady_state(x)
}
\arguments{
\item{x}{A \code{grd_markov} object or transition probability matrix.}
}
\value{
A numeric vector.
}
\description{
Estimate Stationary Distribution
}
